Shift around where BaaS files are installed and run it from a temp directory
This simplifies the project so that the BaaS support files don't have to be
copied to the bundle (which SPM makes very awkward) and installs everything to
a small set of known locations rather than RealmServer having to know things
like the version of mongo being used.

Running mongo and baas out of a new temp directory each time fixes the problem
where sometimes things got into a broken state and the build directory had to
be blown away.