Refactor parts of RLMQueryUtil to reduce template instantiations (#7005)
The main change here is to transform queries which have a constant value on the
LHS and a keypath on the RHS into one with the keypath on the RHS when
possible. This eliminates a large number of redundant template instantiations,
reducing the library size a bit and improving the build time. It also happens
to make some queries which previous threw an exception work.

C++17 lets us replace some of the template metaprogramming with `if constexpr`
making things a bit simpler.

Exception throwing has been shuffled around to reduce the size of the exception
handler tables, and a few things have been refactored to cut down on unneeded
duplication due to tmeplating.

In total this bumps about 100 kB off the macOS release binary and is a small
net reduction in the amount of non-test code.