Upgrade to core 10.5.0
Schema validation for sync realms now distinguishes between auto-discovered
schemas and explicitly set schemas so we need a bunch of awkward logic to set
that properly.

List notifiers now send a notification when the list's parent object is
deleted. We currently don't have a way to report that in the public API, so we
need to explicitly swallow those notifications as otherwise they'll look like
`.iniital` notifications.