Fix exceptions when iterating frozen collections on multiple threads (#7241)
Fast enumeration iterates over a snapshot of the Results rather than the
Results itself so that enumerating objects and mutating them works properly
(we'd otherwise rerun the query after each mutation). For performance, we defer
creating the snapshot until it's needed, i.e. until the Realm enters a write
transaction or is invalidated. To do this, we keep track of all active
RLMFastEnumerator instances on RLMRealm. For frozen Realms, this requires
locking as NSHashTable is not thread-safe and failing to lock access to it
results in corrupting that type's data structures.