Fix running the tests via spm with Xcode 12.5
Loading test bundles with a deployment target older than macOS 11 into xctest
when using SPM and Xcode 12.5 causes crashes when accessing objective-c
metadata, presumably because it's incorrectly expecting the modern format. This
appears to be specific to xctest, so we don't want to bump the minimum
deployment target of everything, and instead only the tests.