Fix crashes in -[PrimitiveArrayPropertyTests testFastEnumeration]
The compiler appears to be miscompiling this function and attempting to unify
all of the `values` arrays into a single one, but what it actually does is read
from an uninitialized local from later in the function instead of the correct
one.

Work around this by sticking each of the subparts of the test in a separate
obj-c block to foil the optimization.