Ensure that the source version is pinned while dispatching changesets to another queue
When using Combine to receive notifications on a different queue, we need to
ensure that the changeset version matches the object version, which requires
pinning that version. TSRs no longer do that pinning, so use a frozen object
instead.

Also do some refactoring that turned out to be unneeded but eliminates some
duplicated code.