Only use the TestHost app for testing on devices
Testing a framework on an iOS device requires using a host application to
launch the tests, but macOS and simulators don't need it. We were using the
test host everywhere as it used to be simpler to do so, but it no longer is and
is causing problems for Catalyst testing. Catalyst seems to sometimes hang for
20+ seconds when running without the GUI visible, which happens when launching
tests using a test host via xcodebuild but not Xcode. Running the tests without
a test host (and so using xctest instead) seems to avoid this problem.