Fix App cleanup between sync tests
SyncManager::reset_for_testing() no longer tears down the network transport due
to the change to hold a network transport instead of a factory, and it turned
out that we were indirectly relying on that to ensure that pending requests
were cancelled.

The main implication of this is that we now need to tear down all App instance
between tests, while previously we used a single App instance for many tests.
This fortunately doesn't appear to have a significant performance impact.