Make Realm async init @MainActor
It currently doesn't appear to be possible for async init to work without being
confined to a global actor. That global actor doesn't need to be MainActor, but
there also doesn't appear to be a way for it to be a user-supplied one and
MainActor is probably the only one people care about for now.