Fix custom obj-c bridging on iOS <14
Versions of iOS before 14 require `as AnyObject` to cast `Any` to a protocol.

Nearly all of these changes are just to re-enable running our tests on older
versions of iOS. Xcode 12.5 had a miscompilation bug when targeting older
versions of iOS, but that seems to have been fixed so we only need the iOS 14
target when using Xcode 12.

Fixes #7604.