Fix using the dynamic API on unmanaged objects before shared schema init (#7790)
If the shared schema was never used to open a Realm, the object schema
instances in it wouldn't have `accessorClass` set, resulting in
`RLMObjectBaseSetObjectForKeyedSubscript()`'s check of `object.class ==
object->_objectSchema.accessorClass` incorrectly concluding that unmanaged
objects were managed. A localized fix for this was possible, but I instead
cleaned up some long-standing tech debt and shifted where `accessorClass` is
populated to a more appropriate location.

We originally captured the *column index* in the blocks used to create the
getters and setters for managed accessor classes. Because column order doesn't
have to match the declared property order, this meant that we had to create
accessor classes after opening a Realm to be able to capture the appropriate
column indices. Early versions of sync broke this: sync originally required
that the column order be the same on all clients, which meant that the sync
algorithm would reorder columns at runtime and we had to deal with that. To
make this work, we switched to capturing the *property index* in the accessor
blocks, and then looked up the column index (and in more recent versions, the
column key) on each property access. This eliminated the need to create
per-Realm accessor classes, but at the time we left the working code alone.

Sync later removed the need for column orders on all clients to match, but the
extra indirection of looking up things by property index turned out to be a
trivial amount of overhead.

This changes it so that we create the accessor classes at schema discovery time
rather than Realm opening time. This fixes the initially reported bug, and
should generally improve performance. When using `objectTypes` it was possible
to create multiple redundant sets of accessor classes, which should no longer
happen. In addition it eliminates some unnecessary copying when using
`objectTypes`.

There is still a follow-up check to ensure that accessor classes are set when
opening a Realm as it can be unset when dynamically creating RLMObjectSchema
instances. This is mostly only applicable to our tests.