Only call the `initialSubscriptions` callback when it should be called
It was called on every Realm open because the `initializationFunction` which
attempted to check if this was the first open was never actually used so
`isFirstOpen` was always `false`, and then the code which checked that value
inverted the condition and did the logic for first open if `isFirstOpen` was
`false` rather than `true`.

The test which attempted to test this did not actually open the Realm a second
time.