Rework sync config initialization
Calculate the path from the partition value when a RLMSyncConfiguration is
constructed rather than when it's set on a RLMRealmConfiguration, and change
RealmSwift.SyncConfiguration to wrap a RLMSyncConfiguration rather than copying
the values out. In combination, these changes make it so that we only calculate
the path when the configuration is constructed, rather than every time a Realm
instance is obtained.