Enable session multiplexing (#8282)
And add an option to disable it. We previously didn't expose any
SyncClientConfig settings, so this includes some refactoring to make
RLMAppConfiguration cover that.

This also fixes an incidental bug where mutating a RLMAppConfiguration after
using it to create a RLMApp could potentially cause problems as we didn't copy
the input object. This was extraordinarily unlikely to ever have been an actual
problem in practice.