Rework the static framework build and consolidate the release packaging
Rather than a separate target, this switches to using a different configuration
for the static framework build. This eliminates most of the duplication and
hacks involved, while unsurprisingly introducing a few new and different problems.

There's now just a single release package rather than separate objc and swift
ones. As we no longer have to deal with the backwards-compatiblity issues from
bitcode, only the obj-c framework built with the most recent version of Xcode
is included rather than one per Xcode version as is done with Swift. The root
directory contains the dynamic xcframework that most people should be using,
and the static xcframework is in the `static` subdirectory. The static
framework now contains slices for all platforms rather than just iOS.