Remove REALM_SWIFT_VERSION logic
This was a legacy of the Swift 1/2/3 migrations where the compiler supported
two different versions of Swift by shipping two different copies of the
frontend built in different modes and we actually built for different Swift
versions. These days we just build once with each Swift compiler version and
don't care about the Swift version, so this logic just complicated things.